diagnosis/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
fastapi>=0.68.0
uvicorn>=0.15.0
Pillow>=8.3.1
numpy>=1.21.0
scikit-learn>=0.24.0
python-multipart>=0.0.5
